# server/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  description: |
    상헌이의 API 창고
  version: '1.0.0'
  title: Sangheon's API Warehouse
  termsOfService: 'http://swagger.io/terms/'
  contact:
    name: Sangheon Kim
    email: ksj8367@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/sangheon-kim/API_Warehouse/1.0.0
  - description: LocalHost
    url: http://localhost:4000
  - description: '개발망'
    url: https://localhost:8000
...
paths:
...
  /projects:
    get:
      tags:
        - RemoteConfiguration
      summary: (Project) Get Project List
      description: Get All Projects List
      operationId: getProjects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SuccessObject'
                properties:
                  _data:
                    type: object
                    properties:
                      projects:
                        type: array
                        items:
                          $ref: '#/components/schemas/Project'
    post:
      tags:
        - RemoteConfiguration
      summary: (Project) Create Project
      description: Create Project
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SuccessObject'
                properties:
                  _data:
                    type: object
                    properties:
                      projects:
                        type: array
                        items:
                          $ref: '#/components/schemas/Project'
  '/projects/{projectId}':
    get:
      tags:
        - RemoteConfiguration
      summary: (Project) Find One Project
      description: Find one project with the projectId
      operationId: getProject
      parameters:
        - name: projectId
          in: path
          description: 프로젝트 ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SuccessObject'
                properties:
                  _data:
                    type: object
                    properties:
                      project:
                        $ref: '#/components/schemas/Project'
        '404':
          description: Error
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ErrorObject'
                properties:
                  status:
                    type: number
                    example: 404
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Project is not exist'
    put:
      tags:
        - RemoteConfiguration
      summary: (Project) Update Project
      description: Update the Project with the ProjectId
      operationId: updateProject
      parameters:
        - name: projectId
          in: path
          description: 프로젝트 ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SuccessObject'
                properties:
                  _data:
                    type: string
                    example: 'OK'
        '404':
          description: No Exist
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ErrorObject'
                properties:
                  status:
                    type: number
                    example: 404
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Project is not exist'
    delete:
      tags:
        - RemoteConfiguration
      summary: (Project) Delete Project
      description: Delete Project By ProjectId
      operationId: deleteProject
      parameters:
        - name: projectId
          in: path
          description: 프로젝트 ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SuccessObject'
                properties:
                  _data:
                    type: string
                    example: 'OK'

        '404':
          description: No Exist
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ErrorObject'
                properties:
                  status:
                    type: number
                    example: 404
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Project is not exist'
  '/projects/{projectId}/items':
    get:
      tags:
        - RemoteConfiguration
      summary: (Item) Get Items
      description: Get Items owned by the project with the project ID
      operationId: getItems
      parameters:
        - name: projectId
          in: path
          description: 프로젝트 ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SuccessObject'
                properties:
                  _data:
                    type: object
                    properties:
                      project:
                        $ref: '#/components/schemas/Project'
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Item'
        '404':
          description: Error
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ErrorObject'
                properties:
                  status:
                    type: number
                    example: 404
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Project is not exist'
    post:
      tags:
        - RemoteConfiguration
      summary: (Item) Create Item
      description: create Items which are owned by Project that has ID which same request parameter.
      operationId: createItem
      parameters:
        - name: projectId
          in: path
          description: 프로젝트 ID
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/Item'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SuccessObject'
                properties:
                  _data:
                    type: object
                    properties:
                      project:
                        $ref: '#/components/schemas/Project'
                      item:
                        $ref: '#/components/schemas/Item'
        '404':
          description: Error
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ErrorObject'
                properties:
                  status:
                    type: number
                    example: 404
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Project is not exist'
...
components:
  schemas:
    ErrorObject:
      type: object
      properties:
        success:
          type: boolean
          description: '성공 여부'
          example: false
        status:
          type: number
          description: 'Status Code'
          example: 500
        error:
          description: 'Error Object'
          type: object
          properties:
            message:
              type: string
              example: 'Something Went wrong'
    SuccessObject:
      type: object
      properties:
        success:
          type: boolean
          description: '성공 여부'
          example: true
        status:
          type: number
          description: 'Status Code'
          example: 200
        _data:
          type: object
          description: 'result'
...
    Project:
      description: Remote Configuration Project Name
      type: object
      properties:
        projectId:
          type: string
          description: 프로젝트 ID
          example: '[프로젝트 ID]'
        description:
          type: string
          description: 해당 프로젝트 설명
          example: '[프로젝트 설명]'
    Item:
      description: Remote Configuration Item Name
      type: object
      properties:
        key:
          type: string
        value:
          type: string
          example: ''
        projectId:
          type: string
...
  requestBodies:
    Item:
      content:
        application/json:
          schema:
            type: object
            properties:
              key:
                type: string
              value:
                type: string
                example: ''
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header